+++ src/animation.ts
